chore(dashboard): organize components and utils
- create pages folder in components
- move helper function getApplicationStatusProps into component/pages/dashboard
- move cards folder into components/pages/dashboard
- relocate types in global into getApplicationStateProps file directly
