+++ apps/ui/src/components/pages/dashboard/ApplicationStatusBar.tsx
