+++ apps/ui/src/components/pages/dashboard/cards/NewApplicationCard.tsx
