--- apps/ui/src/global/types.ts
@@ -1,5 +1,0 @@
-export type ApplicationCardProps = {
-	showEdit: boolean;
-	showActionRequired: boolean;
-	color: string;
-};
